quilkin_xds
Module config

Structs

AwaitingAck

ClientState

ClientTracker

DeltaDiscoveryRes

LocalVersions

Keeps tracking of the local versions of each resource sent from the management server, allowing reconnections to the same/new management servers to send initial versions to reduce the initial response size

Traits

Configuration

Functions

handle_delta_discovery_responses

Processes responses from management servers, applying resources to the proxy

max_grpc_message_size

Returns the configured maximum allowed message size for gRPC messages. When using State Of The World xDS, the message size can get large enough that it can exceed the default limits.

Type Aliases

VersionMap