quilkin_xds::client
Function delta_subscribe

pub async fn delta_subscribe<C: Configuration>(
    config: Arc<C>,
    identifier: String,
    endpoints: Vec<Endpoint>,
    is_healthy: Arc<AtomicBool>,
    notifier: Option<UnboundedSender<String>>,
    resources: &'static [(&'static str, &'static [(&'static str, Vec<String>)])],
) -> Result<JoinHandle<Result<()>>>

Attempts to start a new delta stream to the xDS management server, if the management server does not support delta xDS we return the client as an error