quilkin_xds::config
Function max_grpc_message_size

pub fn max_grpc_message_size() -> usize

Returns the configured maximum allowed message size for gRPC messages. When using State Of The World xDS, the message size can get large enough that it can exceed the default limits.